pyproject.toml
[project]
name = "chatboard"
version = "0.1.0"
description = "A framework for building reliable agents"
authors = [
    { name = "Pavel Schudel", email = "pavel1860@gmail.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
keywords = ["prompt", "llm", "agent", "generation"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy==1.26.4",
    "pydantic>=2.8.2,<3",
    "jsonref==1.1.0",
    "tiktoken==0.5.2",
    "pinecone-text==0.9.0",
    "openai==2.15.0",
    "neo4j==5.28.1",
    "anthropic==0.39.0",
    "langdetect==1.0.9",
    "qdrant-client==1.10.1",
    "docstring_parser==0.16",
    "iso-639==0.4.5",
    "aiohttp==3.10.0",
    "langsmith==0.1.95",
    "asyncpg==0.30.0",
    "tabulate==0.9.0",
    "lxml==6.0.0",
    "PyYAML==6.0.2",
    "PyJWT==2.10.1",
    "google-auth[requests]==2.40.1",
    "google==3.0.0",
    "authlib==1.5.2",
    "pycryptodome==3.23.0",
    "fastapi==0.118.0",
    "google-api-python-client==2.188.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "notebook",
    "nbconvert",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
